// src/app/utilities/enums.ts
export enum constantEnum
{
  pi = "pi",
  e = "e"
}
export enum otherEnums
{
  point = "."
}

export enum operatorEnum
{
  sum = "+" as any,
  subtraction = "–" as any,
  multiplication = "*" as any,
  division = "/" as any,
  power = "^" as any,
  sqrt = "sqrt(" as any,
  module = "%" as any,
  ln = "ln(" as any,
  log = "log(" as any,
}